Proposal: Allow underscores in enum values
Lodash's `upperCase` method will convert all underscores to spaces. So, an enum value of "NOT_CERTIFIED" will get changed to "NOT CERTIFIED" which will not match and cause the lint check to fail. Alternatively, if we use the `toUpper` method, the underscores will be preserved and the lint check will pass.